+++ src/main/java/com/investmetic/domain/strategy/dto/object/StockTypeDto.java
@@ -1,4 +1,4 @@
-package com.investmetic.domain.strategy.dto;
+package com.investmetic.domain.strategy.dto.object;
 
 import com.investmetic.domain.strategy.model.entity.StockType;
 import lombok.Builder;
