+++ src/main/java/com/investmetic/domain/strategy/dto/object/StrategyRegisterRequestDto.java
@@ -1,4 +1,4 @@
-package com.investmetic.domain.strategy.dto;
+package com.investmetic.domain.strategy.dto.object;
 
 import com.investmetic.domain.strategy.model.MinimumInvestmentAmount;
 import com.investmetic.domain.strategy.model.OperationCycle;
